New input obj debug
